requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
mysqlclient==2.2.4
sqlparse==0.5.1
typing-extensions==4.12.2
djangorestframework>=3.14.0